Refactor UI.
The ui.js module was getting out of hand.
Time to properly factor each view into its own module.
