+++ app/js/ui/index.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import AppVC from './AppVC';
+
+export function renderUI() {
+    React.render(<AppVC />, document.getElementById("app"));
+}
